+++ .gitignore
@@ -1,3 +1,4 @@
 token.ini
 .vscode
 __pycache__
+klaybot.ini
