+++ README.md
@@ -1,2 +1,21 @@
 # Klaybot
 My discord bot
+
+
+## Gettings started
+
+Install dependencies with:
+
+```shell
+pip install -r requirements.pip
+```
+
+Create an application for your Discord Bot the [developper portail](https://discordapp.com/developers/applications) then create the associated bot (`Settings > Bot > Add a bot`) and copy the token (See [the official discord walkthrough](https://github.com/reactiflux/discord-irc/wiki/Creating-a-discord-bot-&-getting-a-token)). 
+
+Copy the `klaybot.exemple.ini` to `klaybot.ini` and edit the `discord.token` entry with your own Discord Token.
+
+Launch the bot with:
+
+```shell
+python klaybot.py
+```
