+++ klaybot.exemple.ini
@@ -1,0 +1,2 @@
+[discord]
+token=replace-me
